git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@23167 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/io/comm/FreenetInetAddress.java
@@ -46,7 +46,7 @@ public class FreenetInetAddress {
 			ba = new byte[4];
 			dis.readFully(ba);
 		} else {
-			throw new IOException("Unknown type byte (old form? corrupt stream? too short/long prev field?): "+(int)firstByte);
+			throw new IOException("Unknown type byte (old form? corrupt stream? too short/long prev field?): "+firstByte);
 		}
 		_address = InetAddress.getByAddress(ba);
 		String name = null;
